requirements.txt
Pillow
feedparser
requests
python-dotenv
icalendar
recurring-ical-events
